feat: add Preview feature
